Support `AS` table aliases (6/8)
A "table factor" (name borrowed from the ANSI SQL grammar) is a table
name or a derived table (subquery), followed by an optional `AS` and an
optional alias. (The alias is *not* optional for subqueries, but we
don't enforce that.) It can appear in the FROM/JOIN part of the query.

This commit:
- introduces ASTNode::TableFactor
- changes the parser to populate SQLSelect::relation and Join::relation
  with ASTNode::TableFactor instead of the table name
- changes the parser to only accept subqueries or identifiers, not
  arbitrary expressions in the "table factor" context
- always parses the table name as SQLCompoundIdentifier (whether or not
  it was actually compound).